refactor: composable 함수 분리
